+++ src/main/java/net/replaceitem/discarpet/script/schema/schemas/components/FileUploadSchema.java
@@ -2,10 +2,13 @@ package net.replaceitem.discarpet.script.schema.schemas.components;
 
 import carpet.script.Context;
 import net.dv8tion.jda.api.components.attachmentupload.AttachmentUpload;
+import net.dv8tion.jda.api.interactions.FileType;
 import net.replaceitem.discarpet.script.schema.OptionalField;
 import net.replaceitem.discarpet.script.schema.SchemaClass;
 import net.replaceitem.discarpet.script.schema.SchemaConstructor;
 import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 @SchemaClass(name = "file_upload")
 public class FileUploadSchema implements SchemaConstructor<AttachmentUpload> {
@@ -16,12 +19,14 @@ public class FileUploadSchema implements SchemaConstructor<AttachmentUpload> {
     Integer max_values;
     @OptionalField
     Boolean required = true;
+    @OptionalField
+    List<FileType> file_types = List.of();
 
     @Override
     public AttachmentUpload construct(Context context) {
         AttachmentUpload.Builder builder = AttachmentUpload.create(id);
         if(min_values != null) builder.setMinValues(min_values);
         if(max_values != null) builder.setMaxValues(max_values);
-        return builder.setRequired(required).build();
+        return builder.setFileTypes(file_types).setRequired(required).build();
     }
 }