+++ changelog.md
@@ -1,1 +1,2 @@
 * Added `image_description` and `thumbnail_description` to the embed schema
+* Added `file_types` option for attachment slash command option and file upload component